+++ mephisto/abstractions/blueprints/mixins/onboarding_required.py
@@ -16,9 +16,9 @@ from typing import (
 from mephisto.abstractions.blueprint import BlueprintMixin
 from dataclasses import dataclass, field
 from omegaconf import MISSING, DictConfig
-from mephisto.data_model.qualification import (
+from mephisto.data_model.qualification import QUAL_NOT_EXIST
+from mephisto.utils.qualifications import (
     make_qualification_dict,
-    QUAL_NOT_EXIST,
     find_or_create_qualification,
 )
