+++ mephisto/abstractions/blueprints/mixins/screen_task_required.py
@@ -21,9 +21,9 @@ import types
 from mephisto.abstractions.blueprint import BlueprintMixin
 from dataclasses import dataclass, field
 from omegaconf import MISSING, DictConfig
-from mephisto.data_model.qualification import (
+from mephisto.data_model.qualification import QUAL_NOT_EXIST
+from mephisto.utils.qualifications import (
     make_qualification_dict,
-    QUAL_NOT_EXIST,
     find_or_create_qualification,
 )
