+++ mephisto/abstractions/blueprints/mixins/use_gold_unit.py
@@ -24,9 +24,9 @@ import traceback
 from mephisto.abstractions.blueprint import BlueprintMixin, AgentState
 from dataclasses import dataclass, field
 from omegaconf import MISSING, DictConfig
-from mephisto.data_model.qualification import (
+from mephisto.data_model.qualification import QUAL_NOT_EXIST
+from mephisto.utils.qualifications import (
     make_qualification_dict,
-    QUAL_NOT_EXIST,
     find_or_create_qualification,
 )
 from mephisto.operations.task_launcher import GOLD_UNIT_INDEX
