+++ mephisto/operations/worker_pool.py
@@ -9,8 +9,8 @@ from functools import partial
 from dataclasses import dataclass, fields
 from prometheus_client import Histogram, Gauge, Counter  # type: ignore
 from mephisto.data_model.worker import Worker
-from mephisto.data_model.qualification import worker_is_qualified
 from mephisto.data_model.agent import Agent, OnboardingAgent
+from mephisto.utils.qualifications import worker_is_qualified
 from mephisto.abstractions.blueprint import AgentState
 from mephisto.abstractions.blueprints.mixins.onboarding_required import (
     OnboardingRequired,
