+++ mephisto/scripts/mturk/launch_makeup_hits.py
@@ -23,7 +23,8 @@ from mephisto.data_model.assignment import (
     AssignmentState,
 )
 from mephisto.data_model.unit import Unit
-from mephisto.data_model.qualification import make_qualification_dict, QUAL_EXISTS
+from mephisto.data_model.qualification import QUAL_NOT_EXIST
+from mephisto.utils.qualifications import make_qualification_dict
 from mephisto.operations.task_launcher import COMPENSATION_UNIT_INDEX
 from mephisto.abstractions.providers.mturk.mturk_provider import MTurkProviderArgs
 from mephisto.abstractions.blueprints.mock.mock_blueprint import MockBlueprintArgs
